requirements.txt
streamlit>=1.28.0
streamlit-authenticator>=0.2.3
psycopg2-binary>=2.9.9
pandas>=2.0.0
python-dotenv>=1.0.0
PyYAML>=6.0.1
requests>=2.31.0
sqlalchemy>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
scikit-learn>=1.3.0